+++ tools/go-agent/instrument/plugins/register.go
@@ -21,6 +21,8 @@ import (
 	traceactivation "github.com/apache/skywalking-go/plugin/trace"
 	"github.com/apache/skywalking-go/plugins/core/instrument"
 	"github.com/apache/skywalking-go/plugins/dubbo"
+	fasthttp_client "github.com/apache/skywalking-go/plugins/fasthttp/hostclient"
+	fasthttp_router "github.com/apache/skywalking-go/plugins/fasthttp/router"
 	"github.com/apache/skywalking-go/plugins/gin"
 	goredisv9 "github.com/apache/skywalking-go/plugins/go-redisv9"
 	"github.com/apache/skywalking-go/plugins/go-restfulv3"
@@ -55,6 +57,10 @@ func init() {
 	registerFramework(grpc.NewInstrument())
 	registerFramework(irisv12.NewInstrument())
 	registerFramework(traceactivation.NewInstrument())
+
+	// fasthttp related instruments
+	registerFramework(fasthttp_client.NewInstrument())
+	registerFramework(fasthttp_router.NewInstrument())
 
 	// gorm related instruments
 	registerFramework(gorm_entry.NewInstrument())